pyproject.toml
[project]
name = "pycr2res"
version = "0.1.0"
description = "Python recipes for the cr2res pipeline"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "adari-core>=4.1.0",
    "astropy>=7.1.1",
    "edps>=1.6.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pycpl>=1.0.3",
    "pyesorex>=1.0.3",
    "scipy>=1.16.3",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "ruff>=0.9.0",
]

[tool.uv]
package = false

[[tool.uv.index]]
url = "https://ftp.eso.org/pub/dfs/pipelines/libraries/"
